HP연동 버그 수정 1. fillAmount값 (float)으로 형변환연산 2. Filled Method 수평으로 변경

+++ Assets/PlayerStatusUI.cs
@@ -14,6 +14,6 @@ public class PlayerStatusUI : SingletonMonoBehavior<PlayerStatusUI>
 
     internal void UpdateHP(int hp, int maxHp)
     {
-        hpGuage.fillAmount = hp / maxHp;
+        hpGuage.fillAmount = (float)hp / maxHp;
     }
 }
